refactor and move tests
